// src/server/spider-client/rapport/spider-rapport/prive/secondaire-general/abidjan3/rapport-rentree/interfaces.ts
export interface IChp_A_1 {
  CycleX: string;
  NiveauSerie: string;
  classePedagogique: string;
  classePhysique: string;
  G1: number;
  F1: number;
  T1: number;
  G3: number;
  F3: number;
  T3: number;
}
export interface IChp_B_1 {
  CycleX: string;
  NiveauSerie: string;
  InscritFille: string;
  InscritGarçon: string;
  TotalInscrit: number;
  PresentFille: number;
  PresentGarçon: number;
  TotalPresent: number;
  FilleAdmis: number;
  GarçonAdmis: number;
  TauxGarçonAdmis: number;
  TotalTaux: number;
}

export interface IChp_A_1_3_1 {
  ResponsableEtab: string;
  Discipline: string;
  NomComplet: string;
  Etab: string;
  Contact: string;
}

export interface IChp3_A_1 {
  Genre: string;
  NomComplet: string;
  NomDiplome: string;
  Discipline: string;
  VolumeHoraire: string;
  NumAutEnseigner: string;
  NumCnps: string;
  NomCorps: string;
  Fonction: string;
  RefTypePers: number;
  ProcesVerbale: boolean;
}